reworked project structure, code refactoring, changed setStae tu call with anonimouf functions, corrected misstipes, renamed some functions, change class handling with classNames library, reworked filter options hendling
