+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.3",
     "@testing-library/user-event": "^12.6.2",
+    "classnames": "^2.2.6",
     "date-fns": "^2.16.1",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
