+++ src/assets/filter-options/filter-options.js
@@ -1,0 +1,7 @@
+let filterOptions = {
+    displayAll: "All",
+    displayActive: "Active",
+    displayCompleted: "Completed",
+};
+
+export {filterOptions};
