+++ src/assets/filter-options/index.js
@@ -1,0 +1,3 @@
+import {filterOptions} from './filter-options';
+
+export {filterOptions};
